requirements.txt
tushare>=1.4.29
xtquant>=250516.1.1
pandas>=3.0.2
numpy>=2.4.4
PyMySQL>=1.1.2
python-dotenv>=1.2.2
SQLAlchemy>=2.0.49

Flask>=3.0.0
pyarrow>=18.0.0
akshare>=1.16.0
